[Bronze III] Title: 택시 기하학, Time: 36 ms, Memory: 32540 KB -BaekjoonHub

+++ 백준/Bronze/3053. 택시 기하학/택시 기하학.py
@@ -2,4 +2,4 @@ import sys
 import math 
 n=int(input())
 print(round(n**2*math.pi, 6))
-print("{:.6f}".format(float(2*n*n)))
+print("{:.6f}".format(2*n*n))
